ml_services/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
SQLAlchemy==2.0.23
requests==2.31.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.0.3
Flask-Executor==1.0.0